Move blur to `anomlib.models.components.filters`
