+++ anomalib/models/components/__init__.py
@@ -17,6 +17,7 @@
 from .base import AnomalyModule, DynamicBufferModule
 from .dimensionality_reduction import PCA, SparseRandomProjection
 from .feature_extractors import FeatureExtractor
+from .filters import GaussianBlur2d
 from .sampling import KCenterGreedy
 from .stats import GaussianKDE, MultiVariateGaussian
 
@@ -28,5 +29,6 @@ __all__ = [
     "FeatureExtractor",
     "KCenterGreedy",
     "GaussianKDE",
+    "GaussianBlur2d",
     "MultiVariateGaussian",
 ]
