+++ anomalib/models/components/filters/__init__.py
@@ -1,0 +1,5 @@
+"""Implements filters used by models."""
+
+from .blur import GaussianBlur2d
+
+__all__ = ["GaussianBlur2d"]
