+++ tests/pre_merge/models/components/__init__.py
@@ -1,0 +1,1 @@
+"""Test individual components."""
